refactor(mobile): rename the Kanban tab to Agents
"Kanban" described a board with columns you drag cards between. The
mobile screen is a vertical grouped list -- Working, Needs you, In
review, Ready to merge -- so the name promised an interaction the phone
never had.

"Agents" is what the app already calls these things everywhere else: the
empty state on this very tab reads "No active agents", spawning says
"New agent", and the composer says "Message the agent". The tab title
was the odd one out, not the concept. "Session" stays reserved for the
terminal you open, which is how the rest of the copy already uses it.

The grid icon carried the same board connotation, so it becomes
`activity` -- a pulse, for work in progress.

kanbanView.ts follows to agentsView.ts, matching the <tab>View.ts naming
the other view-models already use (prView, orchestratorView,
notificationView).
