+++ packages/mobile/lib/ProjectPickerSheet.tsx
@@ -6,11 +6,11 @@ import { haptics } from "./haptics";
 import { SHEET_SCROLL_CONTENT, SheetHeader } from "./ui";
 import { useTheme, useThemedStyles } from "./ThemeProvider";
 
-// Picks the active project — the filter behind `useVisibleSessions()` (Kanban +
+// Picks the active project — the filter behind `useVisibleSessions()` (Agents +
 // PRs) and the default project in the spawn screen.
 //
 // The sheet only reports the choice; what happens next is the caller's call.
-// Settings follows a selection through to the Kanban tab, so choosing a project
+// Settings follows a selection through to the Agents tab, so choosing a project
 // and seeing it applied are one step rather than two.
 export const ALL_PROJECTS = "all";
 
@@ -21,7 +21,7 @@ export function ProjectPickerSheet({
 	onSelect,
 	includeAll = true,
 	title = "Active project",
-	subtitle = "Scopes the Kanban board and the PRs tab.",
+	subtitle = "Scopes the Agents and PRs tabs.",
 }: {
 	/** Dismisses the sheet route. */
 	onClose: () => void;
